+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 ## [1.0.5] - Unreleased
+### Improved
+- **Caching** - Cached the app name with app version into the `getAppDetails()` function
 
 ## [1.0.4] - 2026-04-10
 ### Added Languages